feat: add gradient to bg to imitate card table
Here I am tailwind CSS’s basic green colors rather than matching the exact design on the spec. My understanding is that the demands on this project are not to match the figma spec precisely but approximately, with prioritization on semantics usage of tailwind CSS classes. I recognize that adding custom CSS colors to tailwind is an option here, but I am making the assumption that this is not the priority. Calling that out here to note that this choice was taken into consideration.